+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [3.4.1] - 2025-03-16
+
 - (`aaa56b7`) Always allow `ImportExpression`, `SequenceExpression`, and
   `ThisExpression` for `no-top-level-variables`.
 - (`56d7a83`) Allow top-level `WeakMap` and `WeakSet` in recommended config.